refactor: introduce CachedHoardItem
